+++ packages/react-data-grid-addons/src/toolbars/index.ts
@@ -1,2 +1,0 @@
-export { default as AdvancedToolbar } from './AdvancedToolbar';
-export { default as GroupedColumnsPanel } from './GroupedColumnsPanel';