+++ packages/react-data-grid-addons/src/utils.ts
@@ -1,7 +1,0 @@
-export function isEmptyArray(arr: readonly unknown[]): boolean {
-  return arr.length === 0;
-}
-
-export function isEmptyObject<T extends {}>(obj: T): boolean {
-  return Object.keys(obj).length === 0 && obj.constructor === Object;
-}