+++ src/HeaderRow.tsx
@@ -6,14 +6,14 @@ import { assertIsValidKey } from './utils';
 import { DataGridProps } from './DataGrid';
 
 type SharedDataGridProps<R, K extends keyof R, SR> = Pick<DataGridProps<R, K, SR>,
-| 'draggableHeaderCell'
-| 'rows'
-| 'onHeaderDrop'
-| 'onSelectedRowsChange'
-| 'sortColumn'
-| 'sortDirection'
-| 'onSort'
-| 'rowKey'
+  | 'draggableHeaderCell'
+  | 'rows'
+  | 'onHeaderDrop'
+  | 'onSelectedRowsChange'
+  | 'sortColumn'
+  | 'sortDirection'
+  | 'onSort'
+  | 'rowKey'
 >;
 
 export interface HeaderRowProps<R, K extends keyof R, SR> extends SharedDataGridProps<R, K, SR> {
@@ -22,7 +22,7 @@ export interface HeaderRowProps<R, K extends keyof R, SR> extends SharedDataGrid
   lastFrozenColumnIndex: number;
   columns: readonly CalculatedColumn<R, SR>[];
   allRowsSelected: boolean;
-  onColumnResize(column: CalculatedColumn<R, SR>, width: number): void;
+  onColumnResize: (column: CalculatedColumn<R, SR>, width: number) => void;
 }
 
 export default function HeaderRow<R, K extends keyof R, SR>({